Adding support for ChefSpec platform family
When running tests with ChefSpec I load build-essentials recipe and
a number of warnings appear stating that the node's platform is not
supported. The node is the chefspec generated node from Fauxhai.

The node does not have a family at the current version. I am submitting
an additional pull request to FauxHai so that the ChefSpec runner will
use an Ohai response with a platform family