Use most current attribute data for node
Given a recipe like this, the appropriate action will still not run at compile time:
```
node.override['build-essential']['compile_time'] = true
include_recipe 'build-essential'
```

This appears to be due to an issue where the node object referenced in `libraries/timing.rb` is being evaluated before the `node.override` in the recipe itself. Moving the assignment into an attribute file fixes the evaluation order, but it would be nice if this cookbook could still evaluate the node's attribute data based on the freshest node object.

This commit adds the node object as an argument that is passed from the default recipe into the library, no longer relying on the magic of a node variable not defined in the library. As an added bonus, this would allow testing outside of chef, as you could pass a mock node object into the library.