Adding a msys/mingw based build toolchain for Windows
This toolchain is similar to the one used in Rubyinstaller's
DevKit. It is slightly different in that it uses tdb-gcc so
that it can build both 32-bit and 64-bit binaries. It should
be almost identical to the things ksubrama did in omnibus.

There are some idempotence issues with this implementation:
* Once a package has been downloaded, it won't be unpacked
  again after the first try. This means if the first go
  fails, the next chef run won't fix itself