Initial conversion to a custom resource
This just jams everything into a single resource, but that might actually make more sense.  Now that we're guaranteed to be on Chef 12 we can use multipackage too to speed things up

Potentially breaking changes:

Removes support for omnios 151008 which is EOL
Removes warning messaging on Solaris 10. We already don't support Solaris 10 in chef-land so this isn't a real loss